Refactor imports and update types
