+++ src/types.ts
@@ -1,1 +1,7 @@
 export type Mode = 'phrase' | 'word';
+
+export interface Phrase {
+	examples: string[];
+	body: string;
+	meaning: string;
+}
